Properly report size of mailboxes of 2 GibiB or above.
To fix Debian Bug#873668, reported by Andreas Schmidt.
This requires C99's new long long type.

+++ etrn.c
@@ -31,7 +31,7 @@ static int etrn_ok (int sock, char *argbuf)
 }
 
 static int etrn_getrange(int sock, struct query *ctl, const char *id, 
-			 int *countp, int *newp, int *bytes)
+			 int *countp, int *newp, unsigned long long *bytes)
 /* send ETRN and interpret the response */
 {
     int ok, opts;
